requirements.txt
Flask>=2.0
requests
beautifulsoup4
apscheduler
PyYAML
